# dist/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"Usuario": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"nome": {
						"type": "string",
						"default": ""
					},
					"email": {
						"type": "string",
						"default": ""
					},
					"senha": {
						"type": "string",
						"default": ""
					},
					"race": {
						"type": "string",
						"default": ""
					},
					"portadorAneis": {
						"items": {
							"$ref": "#/components/schemas/Anel"
						},
						"type": "array"
					},
					"forjadoPorAneis": {
						"items": {
							"$ref": "#/components/schemas/Anel"
						},
						"type": "array"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Anel": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"nome": {
						"type": "string",
						"default": ""
					},
					"poder": {
						"type": "string",
						"default": ""
					},
					"imagem": {
						"type": "string",
						"default": ""
					},
					"descricao": {
						"type": "string",
						"default": ""
					},
					"portador": {
						"$ref": "#/components/schemas/Usuario"
					},
					"forjadoPor": {
						"$ref": "#/components/schemas/Usuario"
					}
				},
				"required": [
					"id",
					"portador",
					"forjadoPor"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"GeneralResponse__message-string__": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"status": {
						"type": "string"
					},
					"data": {
						"properties": {
							"message": {
								"type": "string"
							}
						},
						"required": [
							"message"
						],
						"type": "object"
					}
				},
				"required": [
					"success",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"GeneralResponse_Anel-Array-or-_message-string__": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"status": {
						"type": "string"
					},
					"data": {
						"anyOf": [
							{
								"items": {
									"$ref": "#/components/schemas/Anel"
								},
								"type": "array"
							},
							{
								"properties": {
									"message": {
										"type": "string"
									}
								},
								"required": [
									"message"
								],
								"type": "object"
							}
						]
					}
				},
				"required": [
					"success",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"GeneralResponse_Anel-or-_message-string__": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"status": {
						"type": "string"
					},
					"data": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/Anel"
							},
							{
								"properties": {
									"message": {
										"type": "string"
									}
								},
								"required": [
									"message"
								],
								"type": "object"
							}
						]
					}
				},
				"required": [
					"success",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GeneralResponse_Anel_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"status": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/Anel"
					}
				},
				"required": [
					"success",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GeneralResponse_number-or-_message-string__": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"status": {
						"type": "string"
					},
					"data": {
						"anyOf": [
							{
								"type": "number",
								"format": "double"
							},
							{
								"properties": {
									"message": {
										"type": "string"
									}
								},
								"required": [
									"message"
								],
								"type": "object"
							}
						]
					}
				},
				"required": [
					"success",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GeneralResponse_null-or-_message-string__": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"status": {
						"type": "string"
					},
					"data": {
						"properties": {
							"message": {
								"type": "string"
							}
						},
						"required": [
							"message"
						],
						"type": "object",
						"nullable": true
					}
				},
				"required": [
					"success",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"BearerAuth": {
				"type": "http",
				"scheme": "bearer"
			}
		}
	},
	"info": {
		"title": "challengedevvo",
		"version": "1.0.0",
		"description": "_One Challenge to rule them all, One Challenge to find them, One Challenge to bring them all, and in the darkness bind them_",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/api/aneis": {
			"get": {
				"operationId": "ListarAneis",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GeneralResponse_Anel-Array-or-_message-string__"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GeneralResponse__message-string__"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"tags": [
					"Anéis"
				],
				"security": [
					{
						"BearerAuth": []
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "CriarAnel",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GeneralResponse_Anel-or-_message-string__"
								}
							}
						}
					},
					"201": {
						"description": "Ring created successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GeneralResponse_Anel_"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					},
					"400": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GeneralResponse__message-string__"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GeneralResponse__message-string__"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"tags": [
					"Anéis"
				],
				"security": [
					{
						"BearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"imagem": {
										"type": "string"
									},
									"forjadoPorId": {
										"type": "number",
										"format": "double"
									},
									"portadorId": {
										"type": "number",
										"format": "double"
									},
									"descricao": {
										"type": "string"
									},
									"poder": {
										"type": "string"
									},
									"nome": {
										"type": "string"
									}
								},
								"required": [
									"imagem",
									"forjadoPorId",
									"portadorId",
									"descricao",
									"poder",
									"nome"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/aneis/{id}": {
			"get": {
				"operationId": "ObterAnelPorId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GeneralResponse_Anel-or-_message-string__"
								}
							}
						}
					},
					"404": {
						"description": "Ring not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GeneralResponse__message-string__"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GeneralResponse__message-string__"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"tags": [
					"Anéis"
				],
				"security": [
					{
						"BearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "AtualizarAnel",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GeneralResponse_Anel-or-_message-string__"
								}
							}
						}
					},
					"400": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GeneralResponse__message-string__"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					},
					"404": {
						"description": "Ring not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GeneralResponse__message-string__"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GeneralResponse__message-string__"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"tags": [
					"Anéis"
				],
				"security": [
					{
						"BearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"imagem": {
										"type": "string"
									},
									"forjadoPorId": {
										"type": "number",
										"format": "double"
									},
									"portadorId": {
										"type": "number",
										"format": "double"
									},
									"descricao": {
										"type": "string"
									},
									"poder": {
										"type": "string"
									},
									"nome": {
										"type": "string"
									}
								},
								"type": "object"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeletarAnel",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GeneralResponse_null-or-_message-string__"
								}
							}
						}
					},
					"404": {
						"description": "Ring not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GeneralResponse__message-string__"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GeneralResponse__message-string__"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"tags": [
					"Anéis"
				],
				"security": [
					{
						"BearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/aneis/upload-image": {
			"post": {
				"operationId": "UploadImage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GeneralResponse_number-or-_message-string__"
								}
							}
						}
					},
					"400": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GeneralResponse__message-string__"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GeneralResponse__message-string__"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"tags": [
					"Anéis"
				],
				"security": [
					{
						"BearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"file": {
										"type": "string",
										"format": "binary"
									}
								},
								"required": [
									"file"
								]
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}